+++ crates/wdl-lint/src/config.rs
@@ -1,6 +1,19 @@
+use std::collections::HashSet;
+
 use serde::Deserialize;
 use serde::Serialize;
 
 /// The configuration for lint rules.
 #[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-pub struct Config {}
+pub struct Config {
+    /// List of keys to ignore in the [`ExpectedRuntimeKeys`] lint.
+    ///
+    /// ## Example
+    ///
+    /// ```toml
+    /// allowed_runtime_keys = ["foo"]
+    /// ```
+    ///
+    /// [`ExpectedRuntimeKeys`]: crate::rules::ExpectedRuntimeKeys
+    pub allowed_runtime_keys: HashSet<String>,
+}